// src/Types/Types.ts
export interface Section {
  Visible?: boolean;
  Name: string;
  ModeLevel: number;
  Description: string;
  Options: KeyValuePair[];
}
export interface ConfData extends ContentType {
  Sections: Section[];
  Autoexec: string;
  AutoexecEnabled?: boolean;
}
// ...
export interface ScriptActions extends ContentType {
  Actions: MenuAction[];
}

// Define the grid container
export interface GameDataList extends ContentType {
  NeedsLogin?: string;
  storeURL?: string;
  Games: GameData[];
}
export interface GameData {
  ID: number;
  Name: string;
  Images: string[];
  ShortName: string;
  SteamClientID: string;
}
// ...
export interface FileData {
  Id: number;
  GameId: number;
  Path: string;
  Content: string;
}
// ...
export interface MenuAction extends ContentType {
  ActionId: string;
  Title: string;
  Type: string;
  InstalledOnly?: boolean;
}

export interface ContentResult<T> {
  Type: string;
  Content: T;
}
export interface ContentType { }
// ...
export interface ExecuteArgs {
  inputData?: string|FileData[]|ConfData;
}